pyproject.toml
[tool.poetry]
name = "reviro-internship"
version = "0.1.0"
description = ""
authors = ["eddy.di <eddy.di.fint@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
pre-commit = "^3.6.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.1"}
factory-boy = "^3.3.0"
faker = "^24.2.0"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
